+++ src/client/sync.rs
@@ -1958,20 +1958,16 @@ pub use crate::subscriptions::SharesChannel;
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
-    use std::vec;
 
     use super::Client;
-    use super::*;
-    use crate::client::mocks::MockGateway;
+    use crate::client::common::tests::*;
     use crate::{connection::ConnectionMetadata, stubs::MessageBusStub};
 
     const CLIENT_ID: i32 = 100;
 
     #[test]
     fn test_connect() {
-        let mut gateway = MockGateway::new();
-
-        let address = gateway.start().expect("Failed to start mock gateway");
+        let (gateway, address) = setup_connect();
 
         let client = Client::connect(&address, CLIENT_ID).expect("Failed to connect");
 
@@ -1984,16 +1980,7 @@ mod tests {
 
     #[test]
     fn test_server_time() {
-        let mut gateway = MockGateway::new();
-
-        let expected_server_time = OffsetDateTime::now_utc().replace_nanosecond(0).unwrap();
-
-        gateway.add_interaction(
-            OutgoingMessages::RequestCurrentTime,
-            vec![format!("49\01\0{}\0", expected_server_time.unix_timestamp())],
-        );
-
-        let address = gateway.start().expect("Failed to start mock gateway");
+        let (gateway, address, expected_server_time) = setup_server_time();
 
         let client = Client::connect(&address, CLIENT_ID).expect("Failed to connect");
